Remove ResetPasswordMain and emailLink code
